Fix bug in Num.h

+++ Number/Num.h
@@ -16,8 +16,8 @@ public:
 	Num();
 	Num(const Num&);
 
-	static uint32_t* bitPatAdd(const uint32_t* opa, const uint32_t* opb, const int len); // Add two unsigned bit pat
-	static uint32_t* bitPatSub(const uint32_t* opa, const uint32_t* opb, const int len); // Sub two bit unsigned bit pat, opa should be larger than opb to make this operation significant
-	static bool bitPatCompare(const uint32_t* opa, const uint32_t* opb, const int len); // Compare two unsinged bit pat, return true if a>=b
+	static uint32_t* bitPatAdd(const uint32_t* opa, const uint32_t* opb, const size_t len); // Add two unsigned bit pat
+	static uint32_t* bitPatSub(const uint32_t* opa, const uint32_t* opb, const size_t len); // Sub two bit unsigned bit pat, opa should be larger than opb to make this operation significant
+	static bool bitPatCompare(const uint32_t* opa, const uint32_t* opb, const size_t len); // Compare two unsinged bit pat, return true if a>=b
 
 };
